submitForm

+++ src/components/Card.js
@@ -2,7 +2,8 @@ export default class Card {
   constructor(
     { cardData, cardSelector, handlePreviewImage, userId },
     handleDeleteButton,
-    myId
+    myId,
+    handleLikeButton
   ) {
     this._name = cardData.name;
     this._link = cardData.link;
@@ -14,7 +15,7 @@ export default class Card {
     this._cardSelector = cardSelector;
     this._handlePreviewImage = handlePreviewImage;
     this._handleDeleteButton = handleDeleteButton;
-    this._likes = cardData.likes;
+    this._handleLikeButton = handleLikeButton;
   }
 
   getId() {
@@ -25,6 +26,25 @@ export default class Card {
     return this._likes.some((like) => like._id === this._myId);
   }
 
+  setLikes(likes) {
+    this._likes = likes;
+    this._renderLikes();
+  }
+
+  _renderLikes() {
+    this._cardLikesCountElement =
+      this._cardElement.querySelector(".card__likes-count");
+    this._cardLikesCountElement.textContent = this._likes.length;
+
+    const cardLikeButton =
+      this._cardElement.querySelector(".card__like-button");
+    if (this.isLiked()) {
+      cardLikeButton.classList.add("card__like-button_active");
+    } else {
+      cardLikeButton.classList.remove("card__like-button_active");
+    }
+  }
+
   _fillCardTemplate() {
     const imageElement = this._cardElement.querySelector(".card__image");
     imageElement.src = this._link;
@@ -32,31 +52,12 @@ export default class Card {
     this._cardElement.id = this._id;
 
     this._cardElement.querySelector(".card__title").textContent = this._name;
-    this._cardLikesCountElement = this._cardElement.querySelector(".card__likes-count");
+    this._cardLikesCountElement =
+      this._cardElement.querySelector(".card__likes-count");
 
     if (this._ownerId !== this._myId) {
       this._cardElement.querySelector(".card__delete-button").remove();
     }
-  }
-
-  setLikes(amountOfLikes) {
-    this._likes = amountOfLikes;
-    this.renderLikes();
-  }
-
-  _renderLikes() {
-    this._cardLikesCountElement =
-      this._cardElement.querySelector(".card__likes-count");
-    this._cardLikesCountElement.textContent =
-      this._likes.length;
-
-      const cardLikeButton = this._cardElement.querySelector(".card__like-button");
-      if (this.isLiked()) {
-        cardLikeButton.classList.add("card__like-button_active");
-      }
-      else {
-        cardLikeButton.classList.remove("card__like-button_active");
-      }
   }
 
   _setEventListeners() {
@@ -65,11 +66,13 @@ export default class Card {
       .addEventListener("click", () => {
         this._toggleLikeButton();
       });
-    this._cardElement
-      .querySelector(".card__delete-button")
-      .addEventListener("click", () => {
-        this._deleteCard(this.getId());
-      });
+    if (this._myId === this._ownerId) {
+      this._cardElement
+        .querySelector(".card__delete-button")
+        .addEventListener("click", () => {
+          this._deleteCard(this.getId());
+        });
+    }
     this._cardElement
       .querySelector(".card__image")
       .addEventListener("click", () => {
@@ -82,7 +85,7 @@ export default class Card {
     this._cardElement = null;
   }
 
-  //handleLikeButton KEEP
+  //handleLikeButton
   _toggleLikeButton() {
     this._cardElement
       .querySelector(".card__like-button")
@@ -90,14 +93,19 @@ export default class Card {
   }
 
   //handleDeleteButton (_deleteCard)
-  _deleteButton() {
-    this._deleteButton = this._cardElement.querySelector(
-      ".card__delete-button"
-    );
-    if (this._userId !== this._owner) {
-      this._deleteButton.classList.add("card__delete-button_hidden");
+  _hideDeleteButton() {
+    this._cardElement.querySelector(".card__delete-button");
+    if (this._ownerId !== this._myId) {
+      this._cardElement.querySelector(".card__delete-button").remove();
     }
   }
+  //   this._deleteButton = this._cardElement.querySelector(
+  //     ".card__delete-button"
+  //   );
+  //   if (this._userId !== this._owner) {
+  //     this._deleteButton.classList.add("card__delete-button-hidden");
+  //   }
+  // }
 
   _getTemplate() {
     const cardElement = document
@@ -111,13 +119,18 @@ export default class Card {
   getView() {
     this._cardElement = this._getTemplate();
     this._setEventListeners();
-    this._cardImage = this._cardElement.querySelector(".card__image");
-    this._cardImage.src = this._link;
-    this._cardImage.alt = this._name;
+    this._cardElement.querySelector(".card__image").src = this._link;
+    this._cardElement.querySelector(".card__image").alt = this._name;
     this._cardElement.querySelector(".card__title").textContent = this._name;
+    this._hideDeleteButton();
     this._renderLikes();
 
     return this._cardElement;
+  }
+
+  remove() {
+    this._cardElement.remove();
+    this._cardElement = null;
   }
 }
